Moved queires to their respective places
